+++ src/main/java/module-info.java
@@ -21,7 +21,6 @@
 module jobj.core.main {
 	requires lombok;
 	requires java.logging;
-	requires org.objenesis;
 	requires java.compiler;
 	exports io.github.astrapi69.check;
 	exports io.github.astrapi69.lang;